% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LGdata.R
\name{LGdata}
\alias{LGdata}
\title{Load Lake George Data}
\usage{
LGdata(data_name)
}
\arguments{
\item{data_name}{A string choosing the data to load.
\tabular{ll}{
\strong{Data name (data_name)} \tab \strong{Data Description} \cr
atm \tab ATM Chemistry \cr
monthly \tab Monthly Lake Level \cr
exo \tab exofull data \cr
ice \tab Ice Cover Levels \cr
}}
}
\description{
Loads data from locally downloaded RData files.
}
\examples{
\dontrun{

#grab secchi data and plot it
ice.levels = LGdata('ice')
}
}
